Continue refactoring `fractal_server/tasks`
