--- fractal_server/tasks/naming.py
@@ -1,19 +1,0 @@
-import re
-
-
-def _normalize_package_name(name: str) -> str:
-    """
-    Implement PyPa specifications for package-name normalization
-
-    The name should be lowercased with all runs of the characters `.`, `-`,
-    or `_` replaced with a single `-` character. This can be implemented in
-    Python with the re module.
-    (https://packaging.python.org/en/latest/specifications/name-normalization)
-
-    Args:
-        name: The non-normalized package name.
-
-    Returns:
-        The normalized package name.
-    """
-    return re.sub(r"[-_.]+", "-", name).lower()
